{{ cookiecutter.repo_name }}/requirements.txt
# deep learning
torch
lightning
wandb
normflows

# snakemake script debugging
typer

# config management
hydra-core

# data science & visualization
matplotlib
numpy
scipy
scikit-learn

# needed for mltools
rich
